Server/Client Config were moved to each CertManager.go
